Consolidate helpers into single helper file. - All the helper modules are available to all the step definitions. - Reduce future code duplication by having all of the modules defined in the same file. - See for docs https://github.com/cucumber/cucumber-ruby/blob/master/features/docs/writing_support_code/world.feature
